signed values

+++ contracts/GeoBache.sol
@@ -5,8 +5,8 @@ pragma experimental ABIEncoderV2;
 contract GeoBache {
   address public owner = msg.sender;
   struct Coords {
-      uint24 lat;
-      uint24 long;
+      int32 lat;
+      int32 long;
   }
   mapping(uint => Coords) public caches;
   Coords[] coordinateList;
@@ -22,7 +22,7 @@ contract GeoBache {
     _;
   }
 
-  function register(uint24 lat, uint24 long) public restricted returns(uint) {
+  function register(int32 lat,int32 long) public restricted returns(uint) {
     Coords memory newCoords = Coords(lat, long);
     coordinateList.push(newCoords);
     caches[coordinateList.length] = newCoords;
@@ -34,9 +34,9 @@ contract GeoBache {
     return caches[index];
   }
 
-  function checkLocation(uint index, uint24 lat, uint24 long) public view returns(bool) {
-    uint24 _lat = caches[index].lat;
-    uint24 _long = caches[index].long;
+  function checkLocation(uint index, int32 lat, int32 long) public view returns(bool) {
+    int32 _lat = caches[index].lat;
+    int32 _long = caches[index].long;
     if(lat >= _lat - 1 && lat <= _lat + 1) {
         if(long >= _long - 1 && long <= _long + 1) {
             return true;
